contextual_agent/requirements.txt
google-adk==1.13.0
a2a-sdk==0.2.5
deprecated>=1.2.14
requests==2.32.4
uvicorn==0.34.0
pydantic>=2.11.3
httpx==0.28.1
click>=8.0.0
python-dotenv>=0.19.0